Fixed build warnings in wxWidgets GUI.

+++ src/wx-utils.cc
@@ -280,7 +280,7 @@ int wx_sendmessage(void* window, int type, INT_PARAM param1, LONG_PARAM param2)
                 ((wxComboBox*) window)->Append((char*) param2);
                 break;
         case WX_CB_SETCURSEL:
-                if (param1 >= 0 && param1 < ((wxComboBox*) window)->GetCount())
+                if (param1 >= 0 && param1 < (int)((wxComboBox*) window)->GetCount())
                         ((wxComboBox*) window)->Select(param1);
                 break;
         case WX_CB_GETCURSEL:
